#28 - Re-organized classes and group methods by output type.
