+++ src/test/java/io/github/avegera/stream/utils/test/CollectionSizeProvider.java
@@ -1,0 +1,23 @@
+package io.github.avegera.stream.utils.test;
+
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.ArgumentsProvider;
+
+import java.util.stream.Stream;
+
+import static org.junit.jupiter.params.provider.Arguments.of;
+
+public class CollectionSizeProvider implements ArgumentsProvider {
+
+    @Override
+    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
+        return Stream.of(
+                of(1),
+                of(3),
+                of(10),
+                of(25),
+                of(100)
+        );
+    }
+}
